// src/app/credit-request/models/dashboardModule.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DashboardModuleType } from './dashboardModuleType';
import { LoansModuleLoan } from './loansModuleLoan';
import { LoansModuleActivityItem } from './loansModuleActivityItem';

// @ts-ignore
/**
 * Module to be displayed in the dashboard. Since is not possible to have the schema generated for the ModuleDetails property, please look   for the schema of the object named with the same name as the ModuleType property
 */
export interface DashboardModule { 
    /**
     * The type of the dashboard module
     */
    moduleType?: DashboardModuleType;
    loans?: Array<LoansModuleLoan>;
    loansActivity?:  { [key: string]: Array<LoansModuleActivityItem>; };
}