// src/app/credit-request/models/dashboardModuleType.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * The type of the dashboard module
 */
export type DashboardModuleType = 'LoansModule';

export const DashboardModuleType = {
    LoansModule: 'LoansModule' as DashboardModuleType
};